Change (de)registration mechanisms, add database columns
Big changes to !register and !unregister mechanisms.

The database now also holds the Discord ID and PUBG nickname.

The registration procedure now first checks if the requesting Discord user already has a link to a PUBG nickname in the database, and if so, tells which one; the one requested or another one. If there is no link, it checks if the PUBG ID that holds the PUBG nickname, asked to be linked to the Discord account, is already linked to another Discord account. It does not tell to whom it belongs. If nothing found, only then a new record will be added to the database.

Before this commit, !unregister required an argument; the PUBG nickname. This was checked against the database and if there was no match, there was no deregistration. However, since the database (now) should only allow Discord and PUBG accounts to occur once, there is no need to specify which account has to be deregistered, because it can only be one. Additionally, the user could have changed their PUBG nickname. On execution of the modified function the (possible outdated) unlinked nickname is returned to the user in the Discord message.